% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{calcLikBranches}
\alias{calcLikBranches}
\title{Calculate likelihood of branches}
\usage{
calcLikBranches(x, log = FALSE)
}
\arguments{
\item{x}{object of class resDating}

\item{log}{whether to return the log of the likelihoods}
}
\description{
Calculate likelihood of branches
}
