% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{calcProbBranches}
\alias{calcProbBranches}
\title{Calculate probability of branches}
\usage{
calcProbBranches(x, log = FALSE)
}
\arguments{
\item{x}{object of class resDating}

\item{log}{whether to return the log of the probability}
}
\description{
Calculate probability of branches
}
