% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{calcResiduals}
\alias{calcResiduals}
\title{Calculate residuals of branches}
\usage{
calcResiduals(x, resample = TRUE)
}
\arguments{
\item{x}{object of class resDating}

\item{resample}{whether to resample or not}
}
\description{
Calculate residuals of branches
}
