% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{DiagnoDating}
\alias{DiagnoDating}
\title{DiagnoDating:  validating dated phylogenies}
\description{
DiagnoDating is a R package to test the validity of dated phylogenies.

The main functions of the package are:
\itemize{

\item TODO
}
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
