% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{estimAlpha}
\alias{estimAlpha}
\title{Estimation of the coalescent rate alpha}
\usage{
estimAlpha(tree, sampling = F)
}
\arguments{
\item{tree}{Dated tree}

\item{sampling}{Whether to sample from the distribution or return the mean}
}
\value{
Estimated value of the coalescent rate alpha
}
\description{
Estimation of the coalescent rate alpha
}
