% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{plot.resDating}
\alias{plot.resDating}
\title{Plotting methods}
\usage{
\method{plot}{resDating}(x, ...)
}
\arguments{
\item{x}{Output from dating}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of results
}
\description{
Plotting methods
}
