% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{plot.resValidate}
\alias{plot.resValidate}
\title{Plotting methods}
\usage{
\method{plot}{resValidate}(x, type = "hist", ...)
}
\arguments{
\item{x}{Output from validate}

\item{type}{Type of plot to do. Currently either 'hist' or 'lastProbBranches' or 'lastResid'.}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of results
}
\description{
Plotting methods
}
