% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{plotLikBranches}
\alias{plotLikBranches}
\title{Plot likelihood of branches}
\usage{
plotLikBranches(x, sub = NA, color = T, minProb = NA, ...)
}
\arguments{
\item{x}{object of class resDating}

\item{sub}{Plot only one of the four subplots}

\item{color}{Whether to use colors to show likelihoods}

\item{minProb}{Minimum likelihood to show as red}

\item{...}{Passed on}
}
\description{
Plot likelihood of branches
}
