% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{plotProbBranches}
\alias{plotProbBranches}
\title{Plot probability of branches}
\usage{
plotProbBranches(x, sub = NA, color = T, minProb = NA, ...)
}
\arguments{
\item{x}{object of class resDating}

\item{sub}{Plot only one of the four subplots}

\item{color}{Whether to use colors to show probabilities}

\item{minProb}{Minimum probability to show as red}

\item{...}{Passed on}
}
\description{
Plot probability of branches
}
