% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{plotResid}
\alias{plotResid}
\title{Plot pseudo-residuals}
\usage{
plotResid(x, sub = NA, ...)
}
\arguments{
\item{x}{object of class resDating}

\item{sub}{Plot only one of the four subplots}

\item{...}{Passed on}
}
\description{
Plot pseudo-residuals
}
