% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{postdistpvals}
\alias{postdistpvals}
\title{Compute posterior distribution of p-values}
\usage{
postdistpvals(x, nrep = 1000, showPlot = F, showTraces = F)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{showPlot}{Whether or not to show the plot}

\item{showTraces}{Whether or not to show traces when running bactdate}
}
\description{
Compute posterior distribution of p-values
}
