% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppcheck.R
\name{ppcheck}
\alias{ppcheck}
\title{Perform posterior predictive check}
\usage{
ppcheck(x, nrep = 1000, showProgress = T, showPlot = F)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{showProgress}{Whether or not to show progress}

\item{showPlot}{Whether or not to show the plot}
}
\description{
Perform posterior predictive check
}
