% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{print.resPDPV}
\alias{print.resPDPV}
\title{Print function for resPDPV objects}
\usage{
\method{print}{resPDPV}(x, ...)
}
\arguments{
\item{x}{output from postdistpval}

\item{...}{Passed on to cat}
}
\value{
Print out details of dating results
}
\description{
Print function for resPDPV objects
}
