% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{print.resValidate}
\alias{print.resValidate}
\title{Print function for resValidate objects}
\usage{
\method{print}{resValidate}(x, ...)
}
\arguments{
\item{x}{output from validate}

\item{...}{Passed on to cat}
}
\value{
Print out details of dating results
}
\description{
Print function for resValidate objects
}
