% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resample}
\alias{resample}
\title{Resample a dated tree using bactdate}
\usage{
resample(x, showProgress = T, showTraces = F, nbIts = 10000)
}
\arguments{
\item{x}{object of class resDating}

\item{showProgress}{Whether or not to show progress}

\item{showTraces}{whether or not to show the MCMC traces}

\item{nbIts}{number of iterations in bactdate resampling}
}
\description{
Resample a dated tree using bactdate
}
