% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDating.R
\name{resDating}
\alias{resDating}
\title{Constructor for the class resDating}
\usage{
resDating(
  dt,
  phy,
  algo = "Unknown",
  model = "poisson",
  rate = 10,
  relax = NA,
  rootdate = NA
)
}
\arguments{
\item{dt}{Dated phylogeny}

\item{phy}{Undated phylogeny (may be rooted or not)}

\item{algo}{Algorithm used for dating}

\item{model}{Model used for dating}

\item{rate}{Clock rate used for dating}

\item{relax}{Relaxation parameter used for dating}

\item{rootdate}{Date of the root inferred by dating}
}
\value{
Object of class resDating
}
\description{
Constructor for the class resDating
}
