% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runBactDating}
\alias{runBactDating}
\title{Date a tree using BactDating}
\usage{
runBactDating(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on to BactDating::bactdate}
}
\value{
resDating object containing results of BactDating analysis
}
\description{
Date a tree using BactDating
}
