% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runLSD}
\alias{runLSD}
\title{Date a tree using LSD}
\usage{
runLSD(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on to Rlsd2::lsd2}
}
\value{
resDating object containing results of LSD analysis
}
\description{
Date a tree using LSD
}
