% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runNodeDating}
\alias{runNodeDating}
\title{Date a tree using node.dating}
\usage{
runNodeDating(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on to ape::node.dating}
}
\value{
resDating object containing results of node.dating analysis
}
\description{
Date a tree using node.dating
}
