% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runRlsd2}
\alias{runRlsd2}
\title{Date a tree using Rlsd2}
\usage{
runRlsd2(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on to Rlsd2::lsd2}
}
\value{
resDating object containing results of LSD analysis
}
\description{
Date a tree using Rlsd2
}
