% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runRTT}
\alias{runRTT}
\title{Run root-to-tip analysis}
\usage{
runRTT(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on to BactDating::roottotip}
}
\value{
resDating object containing results of root-to-tip analysis
}
\description{
Run root-to-tip analysis
}
