% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMaster.R
\name{simMaster}
\alias{simMaster}
\title{Simulation with structure using Master}
\usage{
simMaster(demes = 3, coalrate = 1, migrate = 0.1, samplingDates)
}
\arguments{
\item{demes}{Number of demes}

\item{coalrate}{Within-deme coalescent rate}

\item{migrate}{Migration rate}

\item{samplingDates}{A list containing for each population a vector of sampling dates}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation with structure using Master
}
