% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simStructure.R
\name{simStructure}
\alias{simStructure}
\title{Simulation with structure}
\usage{
simStructure(
  popStarts = c(2020, 2021, 2022),
  globalNeg = 1,
  NeFunPop,
  samplingDates
)
}
\arguments{
\item{popStarts}{Vector of dates when populations start}

\item{globalNeg}{Value of Ne*g for the global population}

\item{NeFunPop}{Function determining the demographic trajectory of populations, default is CaveDive model with N=2 and h=0.2}

\item{samplingDates}{A list containing for each population a vector of sampling dates}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation with structure
}
