% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{takeSample}
\alias{takeSample}
\title{Extract from BactDating output (r) the tree corresponding to a specific iteration in (w)}
\usage{
takeSample(r, w = nrow(r$record))
}
\arguments{
\item{r}{Output from BactDating}

\item{w}{Iteration to extract (default is the last iteration)}
}
\value{
BactDating object corresponding to specified iteration
}
\description{
Extract from BactDating output (r) the tree corresponding to a specific iteration in (w)
}
