% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testImbalance.R
\name{testImbalance}
\alias{testImbalance}
\title{Test to determine whether there are lineages with unbalanced dates
For each node, perform a Kruskal-Wallis test for the dates on the left and on the right}
\usage{
testImbalance(tree, dates)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}
}
\value{
p-values
}
\description{
Test to determine whether there are lineages with unbalanced dates
For each node, perform a Kruskal-Wallis test for the dates on the left and on the right
}
