% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ValidateDating}
\alias{ValidateDating}
\title{ValidateDating: TODO}
\description{
ValidateDating is TODO

The main functions of the package are:
\itemize{

\item TODO
}
}
\author{
Xavier Didelot \email{xavier.didelot@gmail.com}
}
