% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{calcProbBranches}
\alias{calcProbBranches}
\title{Calculate probability of branches}
\usage{
calcProbBranches(x, cumul = FALSE, log = FALSE)
}
\arguments{
\item{x}{object of class resDating}

\item{cumul}{whether to return the cumulative probability}

\item{log}{whether to return the log of the probability}
}
\description{
Calculate probability of branches
}
