% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{calcResiduals}
\alias{calcResiduals}
\title{Calculate normal residuals of branches}
\usage{
calcResiduals(x)
}
\arguments{
\item{x}{object of class resDating}
}
\description{
Calculate normal residuals of branches
}
