% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{postdistpvals}
\alias{postdistpvals}
\title{Compute posterior distribution of p-values}
\usage{
postdistpvals(x, nrep = 1000)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}
}
\description{
Compute posterior distribution of p-values
}
