% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resDating.R
\name{print.resDating}
\alias{print.resDating}
\title{Print function for resDating objects}
\usage{
\method{print}{resDating}(x, ...)
}
\arguments{
\item{x}{output from dating}

\item{...}{Passed on to cat}
}
\value{
Print out details of dating results
}
\description{
Print function for resDating objects
}
