% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runDating}
\alias{runDating}
\title{Run a dating analysis}
\usage{
runDating(tree, dates, algo = "BactDating", rate = NA, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{algo}{Algorithm to use, can be one of: LSD, node.dating, BactDating, treedater, TreeTime}

\item{rate}{If provided, force analysis to use specified value for rate}

\item{...}{Passed on to dating algorithm}
}
\value{
resDating object containing results of dating analysis
}
\description{
Run a dating analysis
}
