% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runLSD}
\alias{runLSD}
\title{Date a tree using LSD}
\usage{
runLSD(tree, dates, rate = NA, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{rate}{If provided, force analysis to use specified value for rate}

\item{...}{Ignored for the time being}
}
\value{
resDating object containing results of LSD analysis
}
\description{
Date a tree using LSD
}
