% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{runTreeDater}
\alias{runTreeDater}
\title{Date a tree using treedater}
\usage{
runTreeDater(tree, dates)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}
}
\value{
dated tree
}
\description{
Date a tree using treedater
}
