% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runTreeDater}
\alias{runTreeDater}
\title{Date a tree using treedater}
\usage{
runTreeDater(tree, dates, rate = NA, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{rate}{If provided, force analysis to use specified value for rate}

\item{...}{Passed on to treedater::dater}
}
\value{
resDating object containing results of treedater analysis
}
\description{
Date a tree using treedater
}
