% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{runTreeTime}
\alias{runTreeTime}
\title{Date a tree using TreeTime}
\usage{
runTreeTime(tree, dates, ...)
}
\arguments{
\item{tree}{Tree to date}

\item{dates}{Dates of leaves in the tree}

\item{...}{Passed on}
}
\value{
resDating object containing results of TreeTime analysis
}
\description{
Date a tree using TreeTime
}
