% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{validate}
\alias{validate}
\title{Check whether a dating analysis is valid}
\usage{
validate(x, nrep = 500, showPlot = T)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{showPlot}{whether or not to show the plot}
}
\description{
Check whether a dating analysis is valid
}
