% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{validate}
\alias{validate}
\title{Check whether a dating analysis is valid}
\usage{
validate(
  x,
  nrep = 1000,
  resampling = 0,
  showProgress = T,
  showTraces = F,
  nbIts = 10000
)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{resampling}{method used for resampling: 0 (default) for using given sample, 1 to use the approximate posterior resampling method, 2 to use BactDating resampling}

\item{showProgress}{Whether or not to show progress}

\item{showTraces}{whether or not to show the MCMC traces}

\item{nbIts}{number of iterations in bactdate resampling}
}
\description{
Check whether a dating analysis is valid
}
