% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{validate}
\alias{validate}
\title{Check whether a dating analysis is valid}
\usage{
validate(x, nrep = 1000, resampling, nstore = 1000, showPlot = T)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{resampling}{whether to use the approximate posterior resampling method, default is to do this iff no posterior sample is available}

\item{nstore}{size of the store to use in approximate posterior resampling method}

\item{showPlot}{whether or not to show the plot}
}
\description{
Check whether a dating analysis is valid
}
