% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{validate}
\alias{validate}
\title{Check whether a dating analysis is valid}
\usage{
validate(
  x,
  nrep = 1000,
  resampling = 0,
  nstore = 1000,
  showProgress = T,
  showPlot = T,
  showLast = F
)
}
\arguments{
\item{x}{object of class resDating}

\item{nrep}{number of repeats to perform}

\item{resampling}{method used for resampling: 0 (default) for using given sample, 1 to use the approximate posterior resampling method, 2 to use BactDating resampling}

\item{nstore}{size of the store to use in approximate posterior resampling method}

\item{showProgress}{Whether or not to show progress}

\item{showPlot}{whether or not to show the plot}

\item{showLast}{whether or not to show the residuals for the last sample}
}
\description{
Check whether a dating analysis is valid
}
